Separate config creation from other activation steps
During activation we have 4 subsequent steps:

1. Acquire the activation lock
2. Create the core config (incl. backup and restore in case of issues)
3. Precompile some more files for the core
4. Restart of the core or reload of the config

This change separates these steps a little more. And makes it possible
to extract the code in the next commit to use it as common code in both
activation implementations.

+++ cmk/base/automations/check_mk.py
@@ -813,58 +813,61 @@ class AutomationRestart(Automation):
     # TODO: Cleanup duplicate code with cmk.base.core.do_restart()
     def execute(self, args: List[str]) -> core_config.ConfigurationWarnings:
         with redirect_stdout(open(os.devnull, "w")):
-            if config.monitoring_core == "nagios":
-                objects_file = cmk.utils.paths.nagios_objects_file
-            else:
-                objects_file = cmk.utils.paths.var_dir + "/core/config"
-
             try:
-                backup_path = None
                 if cmk.base.core.try_get_activation_lock():
                     raise MKAutomationError("Cannot activate changes. "
                                             "Another activation process is currently in progresss")
 
-                if os.path.exists(objects_file):
-                    backup_path = objects_file + ".save"
-                    os.rename(objects_file, backup_path)
-                else:
-                    backup_path = None
-
                 core = create_core()
-                try:
-                    core_config.do_create_config(core, with_agents=True)
-                except Exception as e:
-                    if backup_path:
-                        os.rename(backup_path, objects_file)
-                    if cmk.utils.debug.enabled():
-                        raise
-                    raise MKAutomationError("Error creating configuration: %s" % e)
 
-                if config.monitoring_core == "cmc" or cmk.base.nagios_utils.do_check_nagiosconfig():
-                    if backup_path:
+                try:
+                    backup_path = None
+                    if config.monitoring_core == "nagios":
+                        objects_file = cmk.utils.paths.nagios_objects_file
+                    else:
+                        objects_file = cmk.utils.paths.var_dir + "/core/config"
+
+                    if os.path.exists(objects_file):
+                        backup_path = objects_file + ".save"
+                        os.rename(objects_file, backup_path)
+                    else:
+                        backup_path = None
+
+                    try:
+                        core_config.do_create_config(core, with_agents=True)
+                    except Exception as e:
+                        if backup_path:
+                            os.rename(backup_path, objects_file)
+                        if cmk.utils.debug.enabled():
+                            raise
+                        raise MKAutomationError("Error creating configuration: %s" % e)
+
+                    if config.monitoring_core == "cmc" or cmk.base.nagios_utils.do_check_nagiosconfig(
+                    ):
+                        if backup_path:
+                            os.remove(backup_path)
+                    else:
+                        broken_config_path = "%s/check_mk_objects.cfg.broken" % cmk.utils.paths.tmp_dir
+                        open(broken_config_path,
+                             "w").write(open(cmk.utils.paths.nagios_objects_file).read())
+
+                        if backup_path:
+                            os.rename(backup_path, objects_file)
+                        else:
+                            os.remove(objects_file)
+
+                        raise MKAutomationError(
+                            "Configuration for monitoring core is invalid. Rolling back. "
+                            "The broken file has been copied to \"%s\" for analysis." %
+                            broken_config_path)
+                finally:
+                    if backup_path and os.path.exists(backup_path):
                         os.remove(backup_path)
 
-                    core.precompile()
-
-                    cmk.base.core.do_core_action(self._mode())
-                else:
-                    broken_config_path = "%s/check_mk_objects.cfg.broken" % cmk.utils.paths.tmp_dir
-                    open(broken_config_path,
-                         "w").write(open(cmk.utils.paths.nagios_objects_file).read())
-
-                    if backup_path:
-                        os.rename(backup_path, objects_file)
-                    else:
-                        os.remove(objects_file)
-
-                    raise MKAutomationError(
-                        "Configuration for monitoring core is invalid. Rolling back. "
-                        "The broken file has been copied to \"%s\" for analysis." %
-                        broken_config_path)
+                core.precompile()
+                cmk.base.core.do_core_action(self._mode())
 
             except Exception as e:
-                if backup_path and os.path.exists(backup_path):
-                    os.remove(backup_path)
                 if cmk.utils.debug.enabled():
                     raise
                 raise MKAutomationError(str(e))
